Improvements with symlinks
Essentially works with symlinks -- mostly.  But at least it is functional.  Essentially will not 'cd' into a symlinked "folder" based on a partial for some directories, because technically it's a file, not a folder.  It will still display all the "folders" and allow the User to choose from the numbered menu..... so.  This was a quick work-around. Maybe I will work on it in the future.  But before, it didn't work at all with Symlinks, cuz I never had any. 

Ok, now that I think of it, the way I set up my script to create symlinks for my WSL to bridge between all my WSL Distros and a shared folder on my C:\ creates the exact same structure as my $HOME folder should be.  It appears to make the Folders a file of a few bytes, however I cannot view into these files from my Windows with Notepad etc.  I assume it is a file with the Symlink data that is only viewable from a Linux File System, however, the Linux Filesystem seems to read it as an actual folder with the added inode pointer -e.g  '~/Downloads -> /mnt/c/Users/.../.../Linux-Shared-Folder/~/Downloads'.  The command 'pwd' shows my directory as ~/Downloads even if it doesn't technically exist as a ~/Downloads folder on my WSL Distro.  So it is complicated in the Script as the script is to deal with Folders, but this is technically a file, but it also acts as folder, and appears as a folder with 'pwd', but is also not a folder in 'ls' where it is an 'lrwxrwxrwx'.    

I probably won't be assed to figure this out as there is only so much time in the world, but if I do: the error seems to be with some directories, but does largely work.  It is definitely functional, just a bit annoying.  Can't figure out why it doesn't work for some directories specifically.